pyproject.toml
[tool.poetry]
name = "pythonproject2"
version = "0.1.0"
description = ""
authors = ["CHAMOMILE <CHAMOM2LES@outlook.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
aiogram = "^3.6.0"
sqlalchemy = "^2.0.30"
alembic = "^1.13.1"
python-dotenv = "^1.0.1"
requests = "^2.32.2"
black = "^24.4.2"
asyncpg = "^0.29.0"
aiogram-widgets = "^1.2.7"
pymongo = "^4.7.3"
motor = "^3.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.poe.tasks]
start-bot = "python server.py"
